requirements.txt
python-dotenv
fastapi[standard]
sqlmodel
pytest
pytest-cov
psycopg2-binary
pylint
black